Version 1.1.3
This commit is just as backup.

The app will now show albums! And you can playback songs from the song list or from the album list.

I also fixed a bug when returning from the player view (now the list updates with the correct data)

Need to find a way animate the Toolbar, but remove the excess space left when displaying the Up Navigation Button.