Version 1.1.4
The app should now be fully functional!

The artist and album list will update it's view according to how you interact with the app

The next update should have a better implemented musicRetriever.prepare() method due to the fact that the query might fuck up and the data will not be prepared as expected.

Also I should look up for a solution to the Up Navigation animation problem (animateLayoutChanges for the toolbar layout makes the Up Navigation view to hold space in the toolbar after completing it's fade away animation)